finishing PR review requests
