+++ apps/opik-backend/src/main/java/com/comet/opik/domain/AutomationModelEvaluatorMapper.java
@@ -18,4 +18,6 @@ interface AutomationModelEvaluatorMapper {
     LlmAsJudgeAutomationRuleEvaluatorModel map(AutomationRuleEvaluatorLlmAsJudge dto);
 
     AutomationRuleEvaluatorLlmAsJudge.LlmAsJudgeCode map(LlmAsJudgeAutomationRuleEvaluatorModel.LlmAsJudgeCode detail);
+
+    LlmAsJudgeAutomationRuleEvaluatorModel.LlmAsJudgeCode map(AutomationRuleEvaluatorLlmAsJudge.LlmAsJudgeCode code);
 }
